+++ src/main/java/com/example/wallet_service/repository/TransactionRepository.java
@@ -1,9 +1,13 @@
 package com.example.wallet_service.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.wallet_service.model.Transaction;
 
 public interface TransactionRepository extends JpaRepository<Transaction, String> {
     
+    Optional<Transaction> findByWalletIdAndBetIdAndType(String walletId, String betId, String type);
+
 }